rename mdbt50q board and variant to be consistent
