+++ variants/raytac_mdbt50q_rx/variant.h
